+++ packages/webr/inst/tests-webr/test-webr.R
@@ -35,3 +35,9 @@ webr:::sandbox({
     )
   )
 })
+
+"Emscripten mount and unmount without error"
+webr:::sandbox({
+  webr::mount("/mnt", ".", type = "nodefs")
+  webr::unmount("/mnt")
+})